+++ src/main/java/org/opentripplanner/routing/core/AbstractIntersectionTraversalCostModel.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.routing.core;
 
+import org.opentripplanner.graph_builder.module.osm.WayPropertySetSource.DrivingDirection;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.vertextype.IntersectionVertex;
@@ -13,25 +14,8 @@ public abstract class AbstractIntersectionTraversalCostModel implements
         IntersectionTraversalCostModel {
 
     /** Factor by which absolute turn angles are divided to get turn costs for non-driving scenarios. */
-    protected Double nonDrivingTurnCostFactor = 1.0 / 20.0;
+    protected double nonDrivingTurnCostFactor = 1.0 / 20.0;
 
-    protected Integer minRightTurnAngle = 45;
-    
-    protected Integer maxRightTurnAngle = 135;
-
-    protected Integer minLeftTurnAngle = 225;
-    
-    protected Integer maxLeftTurnAngle = 315;
-
-    /** Returns true if this angle represents a right turn. */
-    protected boolean isRightTurn(int turnAngle) {
-        return (turnAngle >= minRightTurnAngle && turnAngle < maxRightTurnAngle);
-    }
-
-    /** Returns true if this angle represents a left turn. */
-    protected boolean isLeftTurn(int turnAngle) {
-        return (turnAngle >= minLeftTurnAngle && turnAngle < maxLeftTurnAngle);
-    }
 
     /**
      * Computes the turn cost in seconds for non-driving traversal modes.
@@ -67,13 +51,7 @@ public abstract class AbstractIntersectionTraversalCostModel implements
             angleOutOfIntersection += 360;
         }
 
-        int turnAngle = angleOutOfIntersection - angleIntoIntersection;
-
-        if (!options.driveOnRight) {
-            turnAngle = 360 - turnAngle;
-        }
-
-        return turnAngle;
+        return angleOutOfIntersection - angleIntoIntersection;
     }
 
     /* Concrete subclasses must implement this */