+++ src/test/java/org/opentripplanner/graph_builder/module/osm/TestOpenStreetMapGraphBuilder.java
@@ -215,8 +215,7 @@ public class TestOpenStreetMapGraphBuilder extends TestCase {
         assertEquals(wayPropertySet.getDataForWay(way), wayPropertySet.defaultProperties);
 
         // add two equal matches: lane only...
-        OSMSpecifier lane_only = new OSMSpecifier();
-        lane_only.addTag("cycleway", "lane");
+        OSMSpecifier lane_only = new OSMSpecifier("cycleway=lane");
 
         WayProperties lane_is_safer = new WayProperties();
         lane_is_safer.setSafetyFeatures(new P2<Double>(1.5, 1.5));
@@ -224,8 +223,7 @@ public class TestOpenStreetMapGraphBuilder extends TestCase {
         wayPropertySet.addProperties(lane_only, lane_is_safer);
 
         // and footway only
-        OSMSpecifier footway_only = new OSMSpecifier();
-        footway_only.addTag("highway", "footway");
+        OSMSpecifier footway_only = new OSMSpecifier("highway=footway");
 
         WayProperties footways_allow_peds = new WayProperties();
         footways_allow_peds.setPermission(StreetTraversalPermission.PEDESTRIAN);
@@ -237,9 +235,7 @@ public class TestOpenStreetMapGraphBuilder extends TestCase {
         assertEquals(dataForWay, lane_is_safer);
 
         // add a better match
-        OSMSpecifier lane_and_footway = new OSMSpecifier();
-        lane_and_footway.addTag("cycleway", "lane");
-        lane_and_footway.addTag("highway", "footway");
+        OSMSpecifier lane_and_footway = new OSMSpecifier("cycleway=lane;highway=footway");
 
         WayProperties safer_and_peds = new WayProperties();
         safer_and_peds.setSafetyFeatures(new P2<Double>(0.75, 0.75));