+++ src/test/java/org/opentripplanner/routing/core/SimpleTraversalCostModelTest.java
@@ -1,20 +1,21 @@
 package org.opentripplanner.routing.core;
 
-import org.junit.Before;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.LineString;
 import org.opentripplanner.common.geometry.GeometryUtils;
+import org.opentripplanner.graph_builder.module.osm.WayPropertySetSource.DrivingDirection;
+import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.edgetype.StreetTraversalPermission;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.vertextype.IntersectionVertex;
 import org.opentripplanner.routing.vertextype.StreetVertex;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 
 /**
  * Tests for SimpleIntersectionTraversalCostModel.
@@ -31,10 +32,10 @@ public class SimpleTraversalCostModelTest {
 
     public SimpleIntersectionTraversalCostModel costModel;
     
-    @Before
+    @BeforeEach
     public void before() {
         graph = new Graph();
-        costModel = new SimpleIntersectionTraversalCostModel();
+        costModel = new SimpleIntersectionTraversalCostModel(DrivingDirection.RIGHT_HAND_TRAFFIC);
         
         // Initialize the routing request.
         options = new RoutingRequest();
@@ -44,36 +45,78 @@ public class SimpleTraversalCostModelTest {
         options.carAccelerationSpeed  = (2.0);
         options.setStreetSubRequestModes(TraverseModeSet.allModes());
     }
-    
+
     @Test
     public void testCalculateTurnAngle() {
         // Graph for a fictional grid city with turn restrictions
         IntersectionVertex v1 = vertex("maple_1st", new Coordinate(2.0, 2.0), false);
         IntersectionVertex v2 = vertex("maple_2nd", new Coordinate(2.0, 1.0), false);
-        
+        v2.trafficLight = true;
+
         StreetEdge e1 = edge(v1, v2, 1.0, false);
 
         // Edge has same first and last angle.
         assertEquals(90, e1.getInAngle());
         assertEquals(90, e1.getOutAngle());
-        
+
         // 2 new ones
         IntersectionVertex v3 = vertex("test2", new Coordinate(1.0, 1.0), false);
-        
+
         // Second edge
         StreetEdge e2 = edge(v2, v3, 1.0, false);
 
         assertEquals(0, e2.getInAngle());
         assertEquals(0, e2.getOutAngle());
-        
+
         // Difference should be about 90.
         int diff = (e1.getOutAngle() - e2.getInAngle());
         assertEquals(90, diff);
-        
-        int turnAngle = costModel.calculateTurnAngle(e1, e2, options);
-        assertEquals(270, turnAngle);
+
+
+        // calculate the angle for driving on the right hand side
+
+        int rightHandDriveAngle = costModel.calculateTurnAngle(e1, e2, options);
+        assertEquals(270, rightHandDriveAngle);
+        assertTrue(costModel.isTurnAcrossTraffic(rightHandDriveAngle));
+        assertFalse(costModel.isSafeTurn(rightHandDriveAngle));
+
+        // and on the left hand side
+
+        var leftHandDriveCostModel = new SimpleIntersectionTraversalCostModel(DrivingDirection.LEFT_HAND_TRAFFIC);
+        int leftHandDriveAngle = leftHandDriveCostModel.calculateTurnAngle(e1, e2, options);
+        assertEquals(270, leftHandDriveAngle);
+
+        assertTrue(leftHandDriveCostModel.isSafeTurn(leftHandDriveAngle));
+        assertFalse(leftHandDriveCostModel.isTurnAcrossTraffic(leftHandDriveAngle));
+
+        // on a bike the turn cost for crossing traffic (left turn in left hand driving countries)
+        // should be higher than going the opposite direction
+
+        assertEquals(
+                1.6875,
+                costModel.computeTraversalCost(v2, e1, e2, TraverseMode.BICYCLE, options, 40, 40),
+                0.1
+        );
+        assertEquals(
+                0.5625,
+                costModel.computeTraversalCost(v2, e2, e1, TraverseMode.BICYCLE, options, 40, 40),
+                0.1
+        );
+
+        // in left hand driving countries it should be the opposite
+
+        assertEquals(
+                0.5625,
+                leftHandDriveCostModel.computeTraversalCost(v2, e1, e2, TraverseMode.BICYCLE, options, 40, 40),
+                0.1
+        );
+        assertEquals(
+                1.6875,
+                leftHandDriveCostModel.computeTraversalCost(v2, e2, e1, TraverseMode.BICYCLE, options, 40, 40),
+                0.1
+        );
     }
-    
+
     @Test
     public void testTurnDirectionChecking() {
         // 3 points on a roughly on line
@@ -91,8 +134,8 @@ public class SimpleTraversalCostModelTest {
         StreetEdge toEdge = edge(v, w, 1.0, false);
         
         int turnAngle = costModel.calculateTurnAngle(fromEdge, toEdge, options);
-        assertFalse(costModel.isRightTurn(turnAngle));
-        assertFalse(costModel.isLeftTurn(turnAngle));
+        assertFalse(costModel.isSafeTurn(turnAngle));
+        assertFalse(costModel.isTurnAcrossTraffic(turnAngle));
         
         // AKA is a straight ahead.
     }
@@ -199,8 +242,8 @@ public class SimpleTraversalCostModelTest {
         StreetEdge extraEdge = edge(v, u, 1.0, false);
                 
         int turnAngle = costModel.calculateTurnAngle(fromEdge, toEdge, options);
-        assertTrue(costModel.isRightTurn(turnAngle));
-        assertFalse(costModel.isLeftTurn(turnAngle));
+        assertTrue(costModel.isSafeTurn(turnAngle));
+        assertFalse(costModel.isTurnAcrossTraffic(turnAngle));
         
         float fromSpeed = 1.0f;
         float toSpeed = 1.0f;
@@ -232,8 +275,8 @@ public class SimpleTraversalCostModelTest {
         StreetEdge extraEdge = edge(v, u, 1.0, false);
                 
         int turnAngle = costModel.calculateTurnAngle(fromEdge, toEdge, options);
-        assertFalse(costModel.isRightTurn(turnAngle));
-        assertTrue(costModel.isLeftTurn(turnAngle));
+        assertFalse(costModel.isSafeTurn(turnAngle));
+        assertTrue(costModel.isTurnAcrossTraffic(turnAngle));
         
         float fromSpeed = 1.0f;
         float toSpeed = 1.0f;